requirements.txt
requests
python-dotenv
html2text
questionary
InquirerPy